Moved all helper functions to utility files to make main classes cleaner
